Introduction
-Try to use shorter sentences wherever possible

-Separate your thoughts with punctuation

-Avoid being redundant/repetitive. There were a few places where you
repeated some of what you said at the beginning of the sentence at the
end of the sentence. (That’s usually a sign that the sentence has run
on for too long.)

-You tend to use a lot of causal words (“since…”, “because…”, “it
will…”, “can allow…” “causes”), which is fine, but try to be aware of
how many causal connections you make in a single sentence. It’s not
always clear whether you’re saying something of the form:

x —> (y —> z)

or

(x —> y) —> z

It’s a lot harder to tell what the structure of events is when you
describe them in natural language (as opposed to in formal logic). This
is why punctuation is VERY important in writing papers. Think of
punctuation in writing like parentheses in logic.

E.g. (x —> y) —> z

would read like

“since x causes y, it follows that z”

-These are a few sentences I couldn’t make sense of:

“Decoherence creates a dampening affect on the resulting pulse read
from the qubit system when this dampening hits a flat line, decoherence
time, all information is lost from the system.” (paragraph 3)

“The four qubit system is examined in this poster because it lowest
enough system to create encoding codes for encryption and logical
operations in quantum machine that then can be taken advantage of in
five qubit system for the overall decryption process for security
applications.” (last paragraph)

I think the best strategy is just to re-write these. (I think you might
be missing a few words in places, so it’s hard to understand what you
mean.)